refactor: unify domain/auth DTOs under single dto/ package
domain/auth 아래 DTO 패키지가 port/dto(1 file)와 service/dto(10 files)로
나뉘어 있어 Slice Pattern 표준("dto/ 하나만 사용")에 위배되고, 신규
DTO를 추가할 때마다 "port냐 service냐"를 판단해야 하는 마찰이 있었음.

둘을 com.eventitta.domain.auth.dto로 통합:
- port/dto/KakaoUserInfo → dto/KakaoUserInfo
- service/dto/{AuthSessionMetadata, ClientSessionMetadata, KakaoLinkCommand,
  KakaoLoginCommand, LogoutCommand, RefreshCommand, SignInCommand,
  SignUpCommand, SignUpResult, TokenResult} → dto/*

빈 port/dto, service/dto 디렉터리 제거. 타입 시그니처와 public API는
무변경 (패키지만 이동).

영향 범위: domain/auth 내부 서비스 5곳, port(KakaoAuthClient),
infra의 KakaoAuthClientAdapter, api의 AuthController/AuthMapper/
CookieManager/ClientSessionMetadataResolver/UserController 및 각 테스트.
ArchUnit 경계 규칙은 전부 통과.
