+++ eventitta-api/src/main/java/com/eventitta/api/auth/controller/AuthController.java
@@ -14,10 +14,10 @@ import com.eventitta.api.auth.session.ClientSessionMetadataResolver;
 import com.eventitta.api.auth.cookie.CookieManager;
 import com.eventitta.api.auth.oauth.kakao.KakaoAuthorizationSupport;
 import com.eventitta.domain.auth.service.AuthService;
-import com.eventitta.domain.auth.service.dto.ClientSessionMetadata;
-import com.eventitta.domain.auth.service.dto.LogoutCommand;
-import com.eventitta.domain.auth.service.dto.RefreshCommand;
-import com.eventitta.domain.auth.service.dto.TokenResult;
+import com.eventitta.domain.auth.dto.ClientSessionMetadata;
+import com.eventitta.domain.auth.dto.LogoutCommand;
+import com.eventitta.domain.auth.dto.RefreshCommand;
+import com.eventitta.domain.auth.dto.TokenResult;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
