+++ eventitta-api/src/test/java/com/eventitta/api/auth/controller/AuthControllerTest.java
@@ -22,13 +22,13 @@ import com.eventitta.api.auth.oauth.kakao.KakaoAuthorizationSupport;
 import com.eventitta.domain.auth.exception.AuthErrorCode;
 import com.eventitta.domain.auth.repository.RefreshTokenRepository;
 import com.eventitta.domain.auth.service.AuthService;
-import com.eventitta.domain.auth.service.dto.ClientSessionMetadata;
-import com.eventitta.domain.auth.service.dto.KakaoLoginCommand;
-import com.eventitta.domain.auth.service.dto.RefreshCommand;
-import com.eventitta.domain.auth.service.dto.SignInCommand;
-import com.eventitta.domain.auth.service.dto.SignUpCommand;
-import com.eventitta.domain.auth.service.dto.SignUpResult;
-import com.eventitta.domain.auth.service.dto.TokenResult;
+import com.eventitta.domain.auth.dto.ClientSessionMetadata;
+import com.eventitta.domain.auth.dto.KakaoLoginCommand;
+import com.eventitta.domain.auth.dto.RefreshCommand;
+import com.eventitta.domain.auth.dto.SignInCommand;
+import com.eventitta.domain.auth.dto.SignUpCommand;
+import com.eventitta.domain.auth.dto.SignUpResult;
+import com.eventitta.domain.auth.dto.TokenResult;
 import com.eventitta.domain.notification.domain.AlertLevel;
 import com.eventitta.domain.notification.resolver.AlertLevelResolver;
 import com.eventitta.domain.notification.service.AlertNotificationService;
