+++ eventitta-api/src/test/java/com/eventitta/api/user/controller/UserControllerSocialLinkTest.java
@@ -17,8 +17,8 @@ import com.eventitta.api.user.mapper.UserMapper;
 import com.eventitta.domain.auth.exception.AuthErrorCode;
 import com.eventitta.domain.auth.repository.RefreshTokenRepository;
 import com.eventitta.domain.auth.service.AuthService;
-import com.eventitta.domain.auth.service.dto.AuthSessionMetadata;
-import com.eventitta.domain.auth.service.dto.KakaoLinkCommand;
+import com.eventitta.domain.auth.dto.AuthSessionMetadata;
+import com.eventitta.domain.auth.dto.KakaoLinkCommand;
 import com.eventitta.domain.gamification.api.internal.facade.GamificationQueryFacade;
 import com.eventitta.domain.notification.domain.AlertLevel;
 import com.eventitta.domain.notification.resolver.AlertLevelResolver;
