+++ eventitta-domain/src/main/java/com/eventitta/domain/auth/dto/LogoutCommand.java
@@ -1,4 +1,4 @@
-package com.eventitta.domain.auth.service.dto;
+package com.eventitta.domain.auth.dto;
 
 public record LogoutCommand(
     String accessToken,
