+++ eventitta-domain/src/main/java/com/eventitta/domain/auth/service/AuthService.java
@@ -12,19 +12,19 @@ import com.eventitta.domain.auth.port.AuthMailSender;
 import com.eventitta.domain.auth.port.AuthTokenProvider;
 import com.eventitta.domain.auth.port.CredentialAuthenticator;
 import com.eventitta.domain.auth.port.KakaoAuthClient;
-import com.eventitta.domain.auth.port.dto.KakaoUserInfo;
+import com.eventitta.domain.auth.dto.KakaoUserInfo;
 import com.eventitta.domain.auth.repository.AuthIdentityRepository;
 import com.eventitta.domain.auth.repository.RefreshTokenRepository;
-import com.eventitta.domain.auth.service.dto.AuthSessionMetadata;
-import com.eventitta.domain.auth.service.dto.ClientSessionMetadata;
-import com.eventitta.domain.auth.service.dto.KakaoLinkCommand;
-import com.eventitta.domain.auth.service.dto.KakaoLoginCommand;
-import com.eventitta.domain.auth.service.dto.LogoutCommand;
-import com.eventitta.domain.auth.service.dto.RefreshCommand;
-import com.eventitta.domain.auth.service.dto.SignInCommand;
-import com.eventitta.domain.auth.service.dto.SignUpCommand;
-import com.eventitta.domain.auth.service.dto.SignUpResult;
-import com.eventitta.domain.auth.service.dto.TokenResult;
+import com.eventitta.domain.auth.dto.AuthSessionMetadata;
+import com.eventitta.domain.auth.dto.ClientSessionMetadata;
+import com.eventitta.domain.auth.dto.KakaoLinkCommand;
+import com.eventitta.domain.auth.dto.KakaoLoginCommand;
+import com.eventitta.domain.auth.dto.LogoutCommand;
+import com.eventitta.domain.auth.dto.RefreshCommand;
+import com.eventitta.domain.auth.dto.SignInCommand;
+import com.eventitta.domain.auth.dto.SignUpCommand;
+import com.eventitta.domain.auth.dto.SignUpResult;
+import com.eventitta.domain.auth.dto.TokenResult;
 import com.eventitta.domain.user.api.internal.command.RegisterLocalUserCommand;
 import com.eventitta.domain.user.api.internal.command.RegisterSocialUserCommand;
 import com.eventitta.domain.user.api.internal.facade.UserInternalFacade;
