+++ eventitta-domain/src/main/java/com/eventitta/domain/auth/service/RefreshTokenService.java
@@ -3,8 +3,8 @@ package com.eventitta.domain.auth.service;
 import com.eventitta.domain.auth.domain.RefreshToken;
 import com.eventitta.domain.auth.port.AuthTokenProvider;
 import com.eventitta.domain.auth.repository.RefreshTokenRepository;
-import com.eventitta.domain.auth.service.dto.RefreshCommand;
-import com.eventitta.domain.auth.service.dto.TokenResult;
+import com.eventitta.domain.auth.dto.RefreshCommand;
+import com.eventitta.domain.auth.dto.TokenResult;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.Pbkdf2PasswordEncoder;
 import org.springframework.stereotype.Service;
