+++ eventitta-domain/src/main/java/com/eventitta/domain/auth/service/TokenService.java
@@ -3,8 +3,8 @@ package com.eventitta.domain.auth.service;
 import com.eventitta.domain.auth.domain.RefreshToken;
 import com.eventitta.domain.auth.port.AuthTokenProvider;
 import com.eventitta.domain.auth.repository.RefreshTokenRepository;
-import com.eventitta.domain.auth.service.dto.ClientSessionMetadata;
-import com.eventitta.domain.auth.service.dto.TokenResult;
+import com.eventitta.domain.auth.dto.ClientSessionMetadata;
+import com.eventitta.domain.auth.dto.TokenResult;
 import com.eventitta.domain.user.api.internal.facade.UserInternalFacade;
 import com.eventitta.domain.user.api.internal.view.UserAuthView;
 import lombok.RequiredArgsConstructor;
