+++ eventitta-infra/src/test/java/com/eventitta/infra/auth/client/KakaoAuthClientAdapterTest.java
@@ -1,7 +1,7 @@
 package com.eventitta.infra.auth.client;
 
 import com.eventitta.domain.auth.exception.AuthErrorCode;
-import com.eventitta.domain.auth.port.dto.KakaoUserInfo;
+import com.eventitta.domain.auth.dto.KakaoUserInfo;
 import com.eventitta.infra.auth.client.dto.KakaoTokenResponse;
 import com.eventitta.infra.auth.client.dto.KakaoUserResponse;
 import com.eventitta.infra.auth.config.KakaoAuthProperties;
